points-app/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
PyJWT
python-dotenv
email-validator
